demo-python/code/multitenant-rag/requirements.txt
git+https://github.com/azure-samples/ms-identity-python-utilities
aiohttp==3.8.4
aiosignal==1.3.1
artifacts-keyring==0.3.3
async-timeout==4.0.2
attrs==23.1.0
azure-common==1.1.28
azure-core==1.28.0
azure-identity==1.13.0
azure-search-documents==11.4.0a20230509004
bidict==0.22.1
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
cryptography==41.0.1
dataclasses-json==0.5.9
dnspython==2.3.0
eventlet==0.33.3
Flask==2.3.2
Flask-Session==0.5.0
Flask-SocketIO==5.3.4
frozenlist==1.3.3
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
keyring==24.2.0
langchain==0.0.230
langchainplus-sdk==0.0.20
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
ms-identity-web==0.16.6
msal==1.22.0
msal-extensions==1.0.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
portalocker==2.7.0
pycparser==2.21
pydantic==1.10.11
PyJWT==2.7.0
pyOpenSSL==23.2.0
pypdf==3.12.1
python-engineio==4.5.1
python-socketio==5.8.0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
simple-websocket==0.10.1
six==1.16.0
SQLAlchemy==2.0.18
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.0
zope.event==5.0
zope.interface==6.0
